Trying to find out the state after running the step.
Waiting for the completion of the tekton deployment is tricky since we can only observe some side-effects. We can run `kubectl get Kustomization -n flux-system -w` to check if the customization ran successfully BUT this is only feasable after the CRD itself is created. When running the command too early it will just return an error stating this.